feat: add platform-aware checkout init
