+++ storefronts/utils/platformReady.js
@@ -1,0 +1,8 @@
+export async function platformReady() {
+  if (typeof document === 'undefined') return;
+  if (document.readyState === 'loading') {
+    await new Promise(resolve => {
+      document.addEventListener('DOMContentLoaded', resolve, { once: true });
+    });
+  }
+}
